resort data privaty filter, some code improvements, remove sentry integration
