+++ package.json
@@ -43,7 +43,6 @@
     "@feathersjs/express": "^4.3.4",
     "@feathersjs/feathers": "^4.3.4",
     "@feathersjs/socketio": "^4.3.4",
-    "@sentry/integrations": "^5.8.0",
     "@sentry/node": "^5.9.0",
     "ajv": "^6.10.2",
     "async": "^2.6.3",
