+++ src/logger/systemLogger.js
@@ -32,55 +32,12 @@ const systemLogger = winston.createLogger({
 	exitOnError: false,
 });
 
-const secretDataKeys = [
-	'password',
-	'passwort',
-	'new_password',
-	'new-password',
-	'oauth-password',
-	'current-password',
-	'passwort_1',
-	'passwort_2',
-	'password_1',
-	'password_2',
-	'password_verification',
-	'password_control',
-	'PASSWORD_HASH',
-	'password_new',
-];
-const filter = (data) => {
-	const newData = Object.assign({}, data);
-	Object.keys(newData).forEach((key) => {
-		if (secretDataKeys.includes(key)) {
-			delete newData[key];
-		}
-	});
-	return newData;
-};
-
-const secretQueryKeys = [
-	'accessToken',
-	'access_token',
-];
-const filterQuery = (url) => {
-	let newUrl = url;
-	secretQueryKeys.forEach((key) => {
-		if (newUrl.includes(key)) {
-			// first step cut complet query
-			// maybe todo later add query as object of keys and remove keys with filter
-			newUrl = url.split('?')[0];
-			newUrl += '?<secretQuery>';
-		}
-	});
-	return newUrl;
-};
-
 const requestError = (req, userId = 'noUserId', error) => systemLogger.requestError(util.inspect({
 	type: 'RequestError',
 	requestId: req.headers.requestId,
 	userId,
-	url: filterQuery(req.url),
-	data: filter(req.body),
+	url: req.url,
+	data: req.body,
 	method: req.method,
 	timestamp: new Date(),
 	code: error.code,
