+++ src/middleware/index.js
@@ -3,5 +3,5 @@ const addLogger = require('./logger');
 
 module.exports = (app) => {
 	app.configure(addLogger);
-	app.use(notFound()); // @deprecated?
+	// app.use(notFound); // @deprecated?
 };
