+++ src/middleware/not-found-handler.js
@@ -1,7 +1,5 @@
-const errors = require('@feathersjs/errors');
-
-module.exports = function () {
-	return function (req, res, next) {
-		next(new errors.NotFound('Page not found'));
-	};
+const { NotFound } = require('@feathersjs/errors');
+// todo do not work well with error handling, it is produce a 404 and pass it
+module.exports = (req, res, next) => {
+	next(new NotFound('Page not found'));
 };
